backend/pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [{ name = "climateconnect-authors" }]
dependencies = [
    "wheel",
    "aioredis>=1.3.1",
    "amqp>=5.0.9",
    "asgiref>=3.5.0",
    "async-timeout>=3.0.1",
    "attrs>=22.1.0",
    "autobahn>=20.12.3",
    "Automat>=20.2.0",
    "azure-common>=1.1.25",
    "azure-storage-blob>=2.1.0",
    "azure-storage-common>=2.1.0",
    "billiard>=3.6.4.0",
    "black>=23.3.0",
    "cachetools>=4.1.0",
    "celery>=5.2.7",
    "certifi>=2020.4.5.1",
    "cffi>=1.15.0",
    "channels>=3.0.5",
    "channels-redis>=3.4.1",
    "chardet>=3.0.4",
    "charset-normalizer>=2.0.7",
    "click>=8.1.3",
    "click-didyoumean>=0.0.3",
    "click-plugins>=1.1.1",
    "click-repl>=0.2.0",
    "constantly>=15.1.0",
    "cryptography>=38.0.3",
    "daphne>=3.0.2",
    "decorator>=5.1.1",
    "Django>=3.2.18",
    "django-celery-beat>=2.3.0",
    "django-cors-headers>=3.11.0",
    "django-debug-toolbar>=3.5.0",
    "django-filter>=2.4.0",
    "django-redis>=5.1.0",
    "django-rest-knox>=4.2.0",
    "django-storages>=1.9.1",
    "django-timezone-field>=5.0",
    "djangorestframework>=3.13.1",
    "dynaconf>=3.1.8",
    "entrypoints>=0.4",
    "factory-boy>=2.12.0",
    "Faker>=4.0.2",
    "flake8>=4.0.1",
    "h11>=0.11.0",
    "hiredis>=1.1.0",
    "httptools>=0.5.0",
    "hvac>=0.11.2",
    "hyperlink>=20.0.1",
    "idna>=3.3",
    "incremental>=21.3.0",
    "install>=1.3.5",
    "kombu>=5.2.4",
    "loguru>=0.6.0",
    "mailjet-rest>=1.3.4",
    "mccabe>=0.6.1",
    "msgpack>=1.0.4",
    "mypy-extensions>=0.4.3",
    "numpy>=1.23.4",
    "pandas>=1.3.5",
    "pathspec>=0.9.0",
    "Pillow>=9.3.0",
    "platformdirs>=2.5.4",
    "prompt-toolkit>=3.0.14",
    "protobuf>=3.18.3",
    "psycopg2-binary>=2.9.5",
    "pyasn1>=0.4.8",
    "pyasn1-modules>=0.2.8",
    "pycodestyle>=2.8.0",
    "pycparser>=2.20",
    "pyflakes>=2.4.0",
    "PyHamcrest>=2.0.4",
    "pyOpenSSL>=22.1.0",
    "python-crontab>=2.6.0",
    "python-dateutil>=2.8.1",
    "python-dotenv>=0.20.0",
    "python-http-client>=3.2.7",
    "pytz>=2021.3",
    "PyYAML>=6.0",
    "redis>=3.5.3",
    "requests>=2.28.1",
    "rsa>=4.7",
    "sentry-sdk>=1.9.10",
    "service-identity>=18.1.0",
    "shortuuid>=1.0.8",
    "six>=1.14.0",
    "sqlparse>=0.3.1",
    "starkbank-ecdsa>=2.0.3",
    "text-unidecode>=1.3",
    "tomli>=1.2.3",
    "Twisted>=22.10.0",
    "txaio>=20.4.1",
    "typing-extensions>=4.3.0",
    "tzdata>=2022.1",
    "urllib3>=1.26.12",
    "uvicorn>=0.19.0",
    "uvloop>=0.17.0",
    "validators>=0.20.0",
    "vine>=5.0.0",
    "watchgod>=0.6",
    "wcwidth>=0.2.5",
    "websockets>=9.1",
    "zope-interface>=5.1.0",
    "gunicorn>=20.1.0",
    "django-stubs[compatible-mypy]>=4.2.0",
    "setuptools>=67.8.0",
]
requires-python = ">=3.8"
license = { text = "AGPL" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.0.269",
    "django-debug-toolbar>=3.5.0",
    "mypy>=1.2.0",
    "pandas-stubs>=2.0.1.230501",
    "types-Pillow>=9.5.0.4",
    "types-psycopg2>=2.9.21.10",
    "types-python-dateutil>=2.8.19.13",
    "types-redis>=4.5.5.2",
    "types-requests>=2.30.0.0",
    "types-setuptools>=67.7.0.3",
    "types-urllib3>=1.26.25.13",
    "djangorestframework-stubs[compatible-mypy]>=3.14.0",
    "django-filter-stubs>=0.1.3",
    "types-factory-boy>=0.4.1",
    "celery-types>=0.15.0",
]

[tool.ruff]
extend-exclude = ['*/migrations']
ignore = [
    'E501', # since black does not format all long lines (e.g. comments, disable this lint)
    'F541' # f-strings without placeholder
]
# Ignore `F401` (import violations) in all `__init__.py` files, and in `path/to/file.py`.
[tool.ruff.per-file-ignores]
"**/__init__.py" = ["F401"]

[tool.black]
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
# ^/foo.py  # exclude a file named foo.py in the root of the project (in addition to the defaults)
^.*\b(migrations)\b.*$
'''

[tool.mypy]
strict = false # todo: set this to true
implicit_reexport = true # todo: set this to false
mypy_path = ['.']
plugins = ['mypy_django_plugin.main', 'mypy_drf_plugin.main']

[tool.django-stubs]
django_settings_module = "climateconnect_main.settings"

